BankingServer/requirements.txt
asgiref==3.7.2
audioread==3.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
Django==4.2.2
djangorestframework==3.14.0
drf-yasg==1.21.7
fonttools==4.42.1
idna==3.4
importlib-resources==6.0.1
inflection==0.5.1
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.40.1
matplotlib==3.7.2
msgpack==1.0.5
numba==0.57.1
numpy==1.24.2
packaging==23.1
Pillow==9.4.0
platformdirs==3.10.0
pooch==1.7.0
pycparser==2.21
pydub==0.25.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.28.2
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
soundfile==0.12.1
soxr==0.3.6
sqlparse==0.4.4
threadpoolctl==3.2.0
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
zipp==3.16.2
